登録と編集処理をフック化した
